flameguard_project/requirements.txt
asttokens==3.0.0
blinker==1.9.0
catboost==1.2.8
click==8.2.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
executing==2.2.0
Flask==3.1.1
fonttools==4.59.0
graphviz==0.21
imbalanced-learn==0.13.0
imblearn==0.0
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
lightgbm==4.6.0
MarkupSafe==3.0.2
matplotlib==3.10.5
matplotlib-inline==0.1.7
narwhals==2.1.0
nest-asyncio==1.6.0
numpy==2.3.2
packaging==25.0
pandas==2.3.1
parso==0.8.4
pillow==11.3.0
platformdirs==4.3.8
plotly==6.2.0
prompt_toolkit==3.0.51
psutil==7.0.0
pure_eval==0.2.3
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==311
pyzmq==27.0.1
scikit-learn==1.6.1
scipy==1.16.1
six==1.17.0
sklearn-compat==0.1.3
stack-data==0.6.3
threadpoolctl==3.6.0
tornado==6.5.2
traitlets==5.14.3
tzdata==2025.2
wcwidth==0.2.13
Werkzeug==3.1.3
xgboost==3.0.4
